Only open diagnostic virtual lines on hover
The diagnostic lines feature is nice but it can be annoying when there
is a lot of lines for a line and you're making changes on another
(looking at you, rust).

Similar to the previous autocmd for CursorHold, this time we enable
virtual_lines, if the cursor moves we disable it. This can lead to some
open and close behavior if moving around a single line but it's not as
disruptive.